pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastvideo"
version = "1.2.0"
description = "FastVideo"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "torch==2.7.1",
    "torchvision==0.22.1",
    "torchaudio==2.7.1",
    "numpy==1.26.1",
    "einops",
    "tqdm",
    "safetensors",
    "huggingface_hub",
    "opencv-python",
    "pillow",
    "moviepy==1.0.0",
    "decord",
    "open3d",
    "pycolmap==3.11.1",
    "pyceres==2.4",
    "kornia==0.8.1",
    "pyquaternion",
    "plyfile",
    "flow-vis",
    "addict",
    "easydict",
    "ConfigArgParse",
    "PyYAML",
    "python-dotenv",
    "requests",
    "utils3d @ git+https://github.com/EasternJournalist/utils3d.git@d3a577acf0a9ad7e513a1416449a07b6f47d967f",
    "gradio==5.31.0",
    "pydantic==2.10.6",
    "einx",
    "plotly",
    "prettytable",
    "flow_vis",
    "scikit-learn",
    "hydra-core",
    "omegaconf",
    "albumentations",
    "matplotlib",
    "mediapy",
    "scikit-image",
    "timm",
    "PyJWT",
    "gdown",
    "rich",
    "jaxtyping",
    "transformers",
    "accelerate",
    "ftfy",
    "diffusers",
    "lightning",
    "wandb",
    "peft",
    "liger_kernel",
]



[tool.setuptools.packages.find]
exclude = ["assets*", "docker*", "docs", "scripts*"]

[tool.wheel]
exclude = ["assets*", "docker*", "docs", "scripts*"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_calls = true
check_untyped_defs = true
no_implicit_optional = true
